incorporating feedback

+++ projects/boringssl/build.sh
@@ -1,5 +1,6 @@
-#!/bin/bash -eu
-# Copyright 2020 Google Inc.
+#!/bin/bash -eux
+#
+# Copyright 2016 Google Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -47,19 +48,22 @@ for F in $fuzzerFiles; do
   fi
 done
 
-rm -rf genfiles && mkdir genfiles && $SRC/LPM/external.protobuf/bin/protoc asn1_pdu.proto --cpp_out=genfiles --proto_path=$SRC/fuzzing/proto/asn1-pdu-proto/
+if [[ $CFLAGS != *sanitize=memory* ]]; then
+  rm -rf genfiles && mkdir genfiles && $SRC/LPM/external.protobuf/bin/protoc asn1_pdu.proto --cpp_out=genfiles --proto_path=$SRC/fuzzing/proto/asn1-pdu-proto/
 
-fuzzerLPMFiles=$(find $SRC/ -maxdepth 1 -name "*.cc")
+  fuzzerLPMFiles=$(find $SRC/ -maxdepth 1 -name "*.cc")
 
-for F in $fuzzerLPMFiles
-do
-  fuzzerName=$(basename $F .cc)
-  echo "Building fuzzer $fuzzerName"
-  $CXX $CXXFLAGS -I genfiles -I . -I $SRC/libprotobuf-mutator/ -I $SRC/LPM/external.protobuf/include -I include $LIB_FUZZING_ENGINE \
-      $F genfiles/asn1_pdu.pb.cc $SRC/fuzzing/proto/asn1-pdu-proto/asn1_proto_to_der.cc \
-      -I $SRC/boringssl/include ./ssl/libssl.a ./crypto/libcrypto.a \
-      $SRC/LPM/src/libfuzzer/libprotobuf-mutator-libfuzzer.a \
-      $SRC/LPM/src/libprotobuf-mutator.a \
-      $SRC/LPM/external.protobuf/lib/libprotobuf.a \
-      -o $OUT/${fuzzerName} 
-done
+  for F in $fuzzerLPMFiles
+  do
+    fuzzerName=$(basename $F .cc)
+    echo "Building fuzzer $fuzzerName"
+    $CXX $CXXFLAGS -I genfiles -I . -I $SRC/libprotobuf-mutator/ -I $SRC/LPM/external.protobuf/include -I include $LIB_FUZZING_ENGINE \
+        -I $SRC/boringssl/include \
+        $F genfiles/asn1_pdu.pb.cc $SRC/fuzzing/proto/asn1-pdu-proto/asn1_proto_to_der.cc \
+        ./ssl/libssl.a ./crypto/libcrypto.a \
+        $SRC/LPM/src/libfuzzer/libprotobuf-mutator-libfuzzer.a \
+        $SRC/LPM/src/libprotobuf-mutator.a \
+        $SRC/LPM/external.protobuf/lib/libprotobuf.a \
+        -o $OUT/"${fuzzerName}_lpm" 
+  done
+fi
